# perf: improving performance