--- a/internal/parse-options.js
+++ b/internal/parse-options.js
@@ -1,11 +1,17 @@
-// parse out just the options we care about so we always get a consistent
-// obj with keys in a consistent order.
-const opts = ['includePrerelease', 'loose', 'rtl']
-const parseOptions = options =>
-  !options ? {}
-  : typeof options !== 'object' ? { loose: true }
-  : opts.filter(k => options[k]).reduce((o, k) => {
-    o[k] = true
-    return o
-  }, {})
-module.exports = parseOptions
+const parseOptions = options => {
+  if (!options) return {};
+
+  if (typeof options !== 'object') return { loose: true };
+
+  const parsedOptions = {};
+
+  // parse out just the options we care about so we always get a consistent
+  // obj with keys in a consistent order.
+
+  if (options.includePrerelease) parsedOptions.includePrerelease = true;
+  if (options.loose) parsedOptions.loose = true;
+  if (options.rtl) parsedOptions.rtl = true;
+
+  return parsedOptions;
+};
+module.exports = parseOptions;