@jakebailey 2023-03-31
> Just so it doesn't get lost, someone from Tencent wrote this one: [Zzzen/fast-semver](https://github.com/Zzzen/fast-semver) ([twitter.com/callmetsing/status/1641810237404614660](https://twitter.com/callmetsing/status/1641810237404614660)), which is much faster (original semver is 77% slower).

I mentioned this on the Twitter thread, but given it's "only" 4x faster, I really think that an optimized version can be written in JS and be very close. It will just take some benchmarks and profiling to get to the bottom of things. Things like #458 and so on are related here.

Given all of the users downstream appear to usually compare strings, that's the worst case path in the current implementation; it allocates loads of temporary objects to perform the comparisons and then they are thrown away. yarn berry already caches these objects to try and help with this, and I'm considering sending a PR to pnpm to do something similar.

If there were a fast path for string-ly comparison, that might end up being good, though that may be challenging. But, it's pretty straightfoward to fuzz test two implementations for differences.