@jakebailey 2023-03-31
I have a PR I'm going to send later today to pnpm which simply caches `new Range` and then passes in `new SemVer` to `test`, which eliminates the slowdown with minimal memory overhead.

But, I can try and pull this branch in and see if anything changes.