@jakebailey 2023-03-31
Alright, so on pnpm main, this PR actually appears to be quite good!

```
Done in 1m 53.8s
total time:  114.05s
user time:   132.65s
system time: 30.62s
CPU percent: 143%
max memory:  1661 MB
```

Comparing to https://github.com/pnpm/pnpm/pull/6336:

![image](https://user-images.githubusercontent.com/5341706/229239363-e325ce98-0f1d-4b50-8064-f707782fcacb.png)

Let me try the combo of both.