@jakebailey 2023-03-31
The combo of caching + this PR seems to be the fastest yet:

```
Done in 1m 44s
total time:  104.31s
user time:   118.27s
system time: 30.84s
CPU percent: 142%
max memory:  1771 MB
```

So, this PR is a pretty good help for the current code.