@jakebailey 2023-03-31
Of course, it's not as fast as blindly cachinng the world, but that's not the best idea either.