@wraithgar 2023-04-04
Sorry that I missed this, catching up now.  @H4ad has done some very helpful work over on [ssri](https://github.com/npm/ssri/pulls?q=is%3Apr+author%3AH4ad) already and I'm working through the last of them now.  Thanks to all of you who are helping comment on this PR and are doing other performance testing/improvement.

I would humbly make the same suggestion here that I made there, that the PRs be broken up logically by optimization type or code area for easier isolation of reviews.  By all means continue the discussion here if you want to solve problems or come up w/ other solutions before then.