@H4ad 2023-04-06
In case of someone wants to push more perf optimizations in this release, I found two interesting things.

---

Those objects are created every time this method is called but the return is a boolean flag, so we could simply cache those values.

https://github.com/npm/node-semver/blob/f4fa069c9304b847a5083cf77c2ea7c568692ab8/ranges/subset.js#L76-L92

---

Above, a lot of comparisons could be early returned when `true`, this will save some time and we could avoid calling `cmp`.

https://github.com/npm/node-semver/blob/f4fa069c9304b847a5083cf77c2ea7c568692ab8/classes/comparator.js#L100-L125