@wraithgar 2023-04-06
If someone does end up looking at the `classes/comparator.js` they should also be aware of #521, for which the code snippet above is responsible.